+++ src/oracledb/impl/thin/protocol.pyx
@@ -284,11 +284,6 @@ cdef class Protocol(BaseProtocol):
             ProtocolMessage protocol_message
             AuthMessage auth_message
 
-        # check if the protocol version supported by the database is high
-        # enough; if not, reject the connection immediately
-        if self._caps.protocol_version < TNS_VERSION_MIN_ACCEPTED:
-            errors._raise_err(errors.ERR_SERVER_VERSION_NOT_SUPPORTED)
-
         # if we can use OOB, send an urgent message now followed by a reset
         # marker to see if the server understands it
         if self._caps.supports_oob \
@@ -640,11 +635,6 @@ cdef class BaseAsyncProtocol(BaseProtocol):
             FastAuthMessage fast_auth_message
             ProtocolMessage protocol_message
             AuthMessage auth_message
-
-        # check if the protocol version supported by the database is high
-        # enough; if not, reject the connection immediately
-        if self._caps.protocol_version < TNS_VERSION_MIN_ACCEPTED:
-            errors._raise_err(errors.ERR_SERVER_VERSION_NOT_SUPPORTED)
 
         # create the messages that need to be sent to the server
         protocol_message = conn_impl._create_message(ProtocolMessage)