+++ src/oracledb/impl/thin/messages.pyx
@@ -1847,6 +1847,8 @@ cdef class ConnectMessage(Message):
                 buf.read_uint32(&flags)
                 if flags & TNS_ACCEPT_FLAG_FAST_AUTH:
                     buf._caps.supports_fast_auth = True
+                if flags & TNS_ACCEPT_FLAG_HAS_END_OF_REQUEST:
+                    buf._caps.supports_end_of_request = True
             buf._caps._adjust_for_protocol(protocol_version, protocol_options)
             buf._transport._full_packet_size = True
         elif buf._current_packet.packet_type == TNS_PACKET_TYPE_REFUSE:
@@ -1923,7 +1925,6 @@ cdef class DataTypesMessage(Message):
             buf.read_uint16(&conv_data_type)
             if conv_data_type != 0:
                 buf.skip_raw_bytes(4)
-        buf._caps._check_supports_end_of_request()
 
     cdef int _write_message(self, WriteBuffer buf) except -1:
         cdef: