+++ src/oracledb/impl/thin/packet.pyx
@@ -60,8 +60,13 @@ cdef class Packet:
         Returns a boolean indicating if the end of request byte is found at the
         end of the packet.
         """
-        cdef char *ptr = cpython.PyBytes_AS_STRING(self.buf)
-        return ptr[self.packet_size - 1] == TNS_MSG_TYPE_END_OF_REQUEST
+        cdef:
+            uint16_t flags
+            char *ptr
+        ptr = cpython.PyBytes_AS_STRING(self.buf)
+        flags = unpack_uint16(<const char_type*> &ptr[PACKET_HEADER_SIZE],
+                              BYTE_ORDER_MSB)
+        return flags & TNS_DATA_FLAGS_END_OF_REQUEST
 
 
 @cython.final