+++ doc/src/release_notes.rst
@@ -16,6 +16,7 @@ Thin Mode Changes
 #)  Fixed bug when calling :meth:`~Connection.gettype()` with an object type
     name containing ``%ROWTYPE``
     (`issue 304 <https://github.com/oracle/python-oracledb/issues/304>`__).
+#)  Tightened up code looking for the end of a database request.
 #)  Restored error message raised when attempting to connect to Oracle Database
     11g.