+++ doc/src/release_notes.rst
@@ -17,6 +17,8 @@ Thin Mode Changes
     name containing ``%ROWTYPE``
     (`issue 304 <https://github.com/oracle/python-oracledb/issues/304>`__).
 #)  Tightened up code looking for the end of a database request.
+#)  Packet output is now immediately flushed in order to avoid losing output
+    due to buffering when multiple threads are running.
 #)  Restored error message raised when attempting to connect to Oracle Database
     11g.